Fixed Post Motion Event & KillAUra blocking.

+++ src/main/java/net/minecraft/entity/Entity.java
@@ -1245,8 +1245,8 @@ public abstract class Entity implements ICommandSender
             forward = forward * f;
             float f1 = MathHelper.sin(yaw * (float)Math.PI / 180.0F);
             float f2 = MathHelper.cos(yaw * (float)Math.PI / 180.0F);
-            this.motionX += strafe * f2 - forward * f1;
-            this.motionZ += forward * f2 + strafe * f1;
+            this.motionX += (double)(strafe * f2 - forward * f1);
+            this.motionZ += (double)(forward * f2 + strafe * f1);
         }
     }
